fix: use ajv with formats everywhere (#3716)
