+++ src/providers/google/util.ts
@@ -1,16 +1,16 @@
-import Ajv from 'ajv';
 import type { AnySchema } from 'ajv';
 import type { GoogleAuth } from 'google-auth-library';
 import Clone from 'rfdc';
 import { z } from 'zod';
 import logger from '../../logger';
 import { maybeLoadFromExternalFile, renderVarsInObject } from '../../util';
+import { getAjv } from '../../util/json';
 import { getNunjucksEngine } from '../../util/templates';
 import { parseChatPrompt } from '../shared';
 import { VALID_SCHEMA_TYPES } from './types';
 import type { Content, FunctionCall, Part, Tool } from './types';
 
-const ajv = new Ajv();
+const ajv = getAjv();
 // property_ordering is an optional field sometimes present in gemini tool configs, but ajv doesn't know about it.
 // At the moment we will just ignore it, so the is-valid-function-call won't check property field ordering.
 ajv.addKeyword('property_ordering');
